requirements.txt
mysql-connector-python-rf
faker
